♻️Refactor: Components, Pages 폴더 분리
