+++ src/Styles/GlobalStyle.js
@@ -3,7 +3,6 @@ import theme from "./theme";
 
 const GlobalStyle = createGlobalStyle`
 * {
-  /* text-align: center; */
   margin: 0;
   padding: 0;
   box-sizing: border-box;
